feat: add must-gather-image annotation to the CSV
Adds an annotation to the operator CSV so that the CNSA must-gather will be run when
the --all-images option is given to 'oc adm must-gather'.
Exit the bundle generation script on error to avoid silent failures.
Don't specify --use-image-digests to bundle generation because the digests are already
set via the nudges and otherwise bundle generation tries to
get the digest for the must-gather image and fails because this script
runs in a hermetic environment.